Dockerfile based on Alpine
This Dockerfile builds a swipl environment from source based on Alpine and musl, with a resulting image size of about 34 MiB.  By comparison, the current image is about 264 MiB.  The smaller size is due to
* smaller base image (Alpine is 5 MiB)
* efficient cleanup of build packages using Alpines `apk --virtual` option
* recently added support for musl instead of glibc (thanks @JanWielemaker!)

This image builds and installs all modules, and I've tested it successfully running pengines, although it likely needs more testing before making this official.  It may also need a little modification of the `build` configuration file to ensure the same support as the current image.